% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizWithSCE.R
\name{vizWithSCE}
\alias{vizWithSCE}
\title{Visualize the integration of bulk DE results with scRNA-seq datasets}
\usage{
vizWithSCE(dat, which)
}
\arguments{
\item{dat}{the output of integrateWithSingleCell, a list with
results table (res), DESeqDataSet (dds), and SingleCellExperiment (sce)}

\item{which}{which gene to use for the violin plot.
This is represented as a number that corresponds to the lowest p-value
e.g. 1 for the lowest p-value, 2 for the second lowest p-value, etc.}
}
\value{
a violin plot (see Description)
}
\description{
This function assists with visualization of the integration
of bulk DE results tables with pre-processed scRNA-seq datasets available
on Bioconductor. After a user has picked a scRNA-seq dataset
using \code{integrateWithSingleCell} from the DESeq2 package,
they will have a results table, DESeqDataSet, and a SingleCellExperiment.
After annotating the SingleCellExperiment (see vignette), they can
provide that list of object to this function.
This function then produces a violin plot of the expression of
genes of interest from the bulk DE analysis across annotated cell types
in the single-cell dataset.
}
\examples{

\dontrun{
dat <- integrateWithSingleCell(res,dds)
vizWithSCE(dat, which=1)
}

}
\author{
Kwame Forbes

Michael Love
}
