% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizWithSCE.R
\name{vizWithSCE}
\alias{vizWithSCE}
\title{Visualize the integration of bulk DE results with Bioconductor single-cell RNA-seq datasets}
\usage{
vizWithSCE(dat, which)
}
\arguments{
\item{dat}{this is a variable of integrateWithSingleCell}

\item{which}{which gene to use for the violin plot. This is represented as a number that corresponds to the lowest adjusted p-value
e.g. 1 for the lowest adjusted p-value, 2 for the second lowest adjusted p-value, 3 for the third, etc.}
}
\value{
nothing
}
\description{
A package that assists with the visualization of the integration of bulk DE results tables with pre-processed scRNA-seq datasets available on Bioconductor, for downstream visualization tasks. After the user has pick a scRNA-seq dataset from integrateWithSingleCell.... The output of the function is...
}
\examples{

\dontrun{
dat <- integrateWithSingleCell(res,dds)
vizWithSCE(dat, which=1)
}

}
\author{
Michael Love

Kwame Forbes
}
