% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizWithSCE.R
\name{vizWithSCE}
\alias{vizWithSCE}
\title{Visualize the integration of bulk DE results with scRNA-seq datasets}
\usage{
vizWithSCE(dat, which)
}
\arguments{
\item{dat}{the output of integrateWithSingleCell, a list with
results table (res), DESeqDataSet (dds), and SingleCellExperiment (sce)}

\item{which}{which gene to use for the violin plot.
This is represented as a number that corresponds to the lowest p-value
e.g. 1 for the lowest p-value, 2 for the second lowest p-value, etc.}
}
\value{
nothing
}
\description{
A package that assists with the visualization of the integration
of bulk DE results tables with pre-processed scRNA-seq datasets available
on Bioconductor, for downstream visualization tasks. After the user has pick
a scRNA-seq dataset from integrateWithSingleCell....
The output of the function is...
}
\examples{

\dontrun{
dat <- integrateWithSingleCell(res,dds)
vizWithSCE(dat, which=1)
}

}
\author{
Kwame Forbes

Michael Love
}
